customGamble added
